+++ database/migrations/2023_08_10_084716_create_cashouts_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('cashouts', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('amount');
-            $table->bigInteger('charges')->default(0);
+            $table->double('amount',15,2);
+            $table->double('charges',15,2)->default(0);
             $table->string('method');
             $table->string('currency');
             $table->string('details')->nullable();
@@ -25,8 +25,8 @@ return new class extends Migration
             $table->string('receiver_id');
             $table->string('transfer_id')->default("Requested");
             $table->string('admin_message')->nullable();
-            $table->bigInteger('balance_before');
-            $table->bigInteger('balance_after');
+            $table->double('balance_before',15,2);
+            $table->double('balance_after',15,2);
             $table->string('status')->default("Requested");
             $table->timestamps();
         });
